move tbplugins up; remove references to libs/*
